media-cleaner/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
prometheus-client==0.21.0
python-dateutil==2.9.0.post0
